+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Feature] Add EmptyDashboard component. (#801)
 
 ## 27.0.1 - 2017-10-04
 - [Patch] Add close dropdown behavior. (#824)